requirements.txt
transformers==4.38.0
torch==2.2.0
accelerate==0.27.0
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
